Added Codelite project
Setup Codelite for my own usage, including the .project and .workspace files which should allow someone else to add this project to Codelite and have Codelite generate/build the project.

Once I get to the point where I have releases, I'll start including a Makefile so people can build this for themselves without Codelite. But right now I added that to the .gitignore so I don't have a bunch of files always wanting to be added to commits.